+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.0.5] - 2026-01-02
+
+### Changed
+
+- Modified db to handle 3 line address fields of longer length
+- Display address on up to 4 separate lines
+
 ## [4.0.4] - 2025-12-03
 
 ### Changed